requirements.txt
flask==3.0.3
langchain==0.3.0
langchain-openai==0.2.0
python-dotenv==1.0.1